Fix: 파일명 변경 및 판매자 상품 조회 api 변경 #76
