+++ src/components/seller/SellerOrderManager.tsx
@@ -1,6 +1,3 @@
-import { Link } from 'react-router-dom';
-import { useEffect, useState } from 'react';
-
 import {
   TableContainer,
   Table,
@@ -16,31 +13,22 @@ import {
   MenuItem,
   InputLabel,
   FormControl,
-  ToggleButton,
 } from '@mui/material';
-import CheckIcon from '@mui/icons-material/Check';
+import { useEffect, useState } from 'react';
 
 import { api } from '../../api/api';
 import { IProduct, IOrderItem } from '../../type';
 import { CATEGORY, QUALITY, ORDER_STATE } from '../../constants/index';
+import { Link } from 'react-router-dom';
 
-export default function SellerOrderList() {
+export default function SellerOrderManager() {
   const _id = localStorage.getItem('_id');
-
   const [productList, setProductList] = useState<IProduct[]>([]);
   const [sortedProductList, setSortedProductList] = useState<IProduct[]>([]);
   const [sortOrder, setSortOrder] = useState('최신순');
-  const [isShow, setIsShow] = useState(false);
   const [orderList, setOrderList] = useState<IOrderItem[]>([]);
 
-  const [productOrderStates, setProductOrderStates] = useState<
-    Record<string, string>
-  >({});
-
-  useEffect(() => {
-    fetchSellerProduct();
-  }, []);
-
+  // 날짜 변환 함수
   function formatDate(dateString: string) {
     return new Intl.DateTimeFormat('ko-KR', {
       year: 'numeric',
@@ -48,40 +36,35 @@ export default function SellerOrderList() {
       day: 'numeric',
     }).format(new Date(dateString));
   }
-  const fetchSellerProduct = async () => {
-    try {
-      const response = await api.getOrderState();
-      const getMatchItem = response.data.item;
-      setProductList(getMatchItem);
-    } catch (error) {
-      console.log('판매 상품 조회 실패', error);
-    }
-  };
-  console.log('판매상품 리스트', productList);
 
   useEffect(() => {
-    const getOrderState = async () => {
+    const fetchSellerProductData = async () => {
       try {
-        const response = await api.getOrderState();
-        const orderState = response.data.item;
-        const orderStatesMap: Record<string, string> = {};
-        orderState.forEach((orderItem: IOrderItem) => {
-          orderItem.products.forEach((product) => {
-            console.log('orderItem', orderItem.state);
-            orderStatesMap[product.name] = orderItem.state;
-          });
-        });
-
-        setProductOrderStates(orderStatesMap);
-        setOrderList(orderState);
+        const response = await api.getProductList();
+        const getMatchItem = response.data.item.filter(
+          (id: IProduct) => id.seller_id === Number(_id),
+        );
+        setProductList(getMatchItem);
       } catch (error) {
-        console.log('주문상태오류', error);
+        console.log('상품 목록 조회에 실패했습니다', error);
       }
     };
-
-    getOrderState();
+    fetchSellerProductData();
   }, []);
 
+  useEffect(() => {
+    const getOrderProductInfo = async () => {
+      try {
+        const response = await api.getOrderProductInfo();
+        setOrderList(response.data.item);
+      } catch (error) {
+        console.log(error);
+      }
+    };
+    getOrderProductInfo();
+  }, []);
+
+  // SORT 정렬
   useEffect(() => {
     let sorted = [...productList];
     switch (sortOrder) {
@@ -124,25 +107,8 @@ export default function SellerOrderList() {
     );
   }
 
-  const getOrderStateLabel = (productId) => {
-    const orderItem = orderList.find((order) =>
-      order.products.some((product) => product._id === productId),
-    );
-
-    if (orderItem) {
-      const orderStateCode = ORDER_STATE.codes.find(
-        (code) => code.code === orderItem.state,
-      );
-      return orderStateCode ? orderStateCode.value : 'Unknown Order State';
-    }
-
-    return '판매중';
-  };
   return (
     <>
-      <Link to={`/user/${_id}/product-create`}>
-        <Button variant="contained">등록하기</Button>
-      </Link>
       <Box
         sx={{
           display: 'flex',
@@ -174,13 +140,12 @@ export default function SellerOrderList() {
                   등록일자
                   <br /> (등록번호)
                 </TableCell>
-
+                <TableCell align="center">카테고리</TableCell>
                 <TableCell align="center">이미지</TableCell>
                 <TableCell align="center">상품명</TableCell>
-                <TableCell align="center">품질</TableCell>
                 <TableCell align="center">가격</TableCell>
-                <TableCell align="center">주문상태</TableCell>
-                <TableCell align="center">공개여부</TableCell>
+                <TableCell align="center">배송상태</TableCell>
+
                 <TableCell align="center"></TableCell>
               </TableRow>
             </TableHead>
@@ -193,10 +158,21 @@ export default function SellerOrderList() {
                     </Typography>
                     ({rows._id})
                   </TableCell>
-
+                  <TableCell align="center">
+                    {CATEGORY.depth2
+                      .filter(
+                        (category) =>
+                          category.dbCode === rows.extra.category[1],
+                      )
+                      .map((categoryName) => (
+                        <Typography key={categoryName.id} variant="body2">
+                          {categoryName.name}
+                        </Typography>
+                      ))}
+                  </TableCell>
                   <TableCell align="center">
                     <img
-                      src={`${rows.products[0].image.path}`}
+                      src={`${rows.mainImages[0].path}`}
                       alt="main-Image"
                       style={{
                         width: '80px',
@@ -208,58 +184,25 @@ export default function SellerOrderList() {
                   </TableCell>
                   <TableCell align="center">
                     <Link to={`http://localhost:5173/product/${rows._id}`}>
-                      {rows.products[0].name}
+                      {rows.name}
                     </Link>
                   </TableCell>
                   <TableCell align="center">
-                    {rows.products[0].quantity.sort
-                      ? QUALITY.find(
-                          (quality) =>
-                            quality.value === rows.products[0].quantity.sort,
-                        )?.name
-                      : QUALITY.find(
-                          (quality) =>
-                            quality.value === rows.products[0].quantity,
-                        )?.name || 'Unknown Quality'}
+                    {rows.price.toLocaleString()}원
                   </TableCell>
                   <TableCell align="center">
-                    {rows.products[0].price.toLocaleString()}원
+                    {ORDER_STATE.codes.find(
+                      (state) => state.code === orderList[0]?.state,
+                    )?.value || 'Unknown State'}
                   </TableCell>
-                  <TableCell align="center">
-                    {getOrderStateLabel(rows._id) || ''}
-                  </TableCell>
-                  <TableCell align="center">
-                    <ToggleButton
-                      value="check"
-                      selected={isShow}
-                      size={'small'}
-                      onChange={() => {
-                        setIsShow(!rows.show);
-                      }}
-                    >
-                      <CheckIcon />
-                    </ToggleButton>
-                  </TableCell>
-                  <TableCell align="center">
-                    <Box
-                      sx={{
-                        display: 'flex',
-                        flexDirection: 'column',
-                        alignItems: 'center',
-                        gap: 1,
-                      }}
-                    ></Box>
-                  </TableCell>
-                  <TableCell align="center">
-                    <Link
-                      to={`/user/${rows._id}/product-update`}
-                      state={{ productId: `${rows._id}` }}
-                    >
-                      <Button type="button" variant="contained">
-                        수정하기
-                      </Button>
-                    </Link>
-                  </TableCell>
+                  <Box
+                    sx={{
+                      display: 'flex',
+                      flexDirection: 'column',
+                      alignItems: 'center',
+                      gap: 1,
+                    }}
+                  ></Box>
                 </TableRow>
               ))}
             </TableBody>
